Reduce the length of the `MasterSite` class IV (#2763)
- Move image-related functions to the ImageHandler class
- Feedback-based fixes
