+++ src/MasterSite.php
@@ -20,12 +20,6 @@ use WP_Post;
 class MasterSite extends TimberSite
 {
     /**
-     * Credit meta field key
-     *
-     */
-    public const CREDIT_META_FIELD = '_credit_text';
-
-    /**
      * Theme directory
      *
      */
@@ -105,7 +99,6 @@ class MasterSite extends TimberSite
         add_action('wp_enqueue_scripts', [PublicAssets::class, 'enqueue_js']);
         add_filter('safe_style_css', [$this, 'set_custom_allowed_css_properties']);
         add_filter('wp_kses_allowed_html', [$this, 'set_custom_allowed_attributes_filter'], 10, 2);
-        add_action('save_post', [$this, 'set_featured_image'], 10, 2);
         add_filter('wp_insert_post_data', [$this, 'require_post_title'], 10, 1);
         // Save "p4_global_project_tracking_id" on post save.
         add_action('save_post', [$this, 'save_global_project_id'], 10, 1);
@@ -113,11 +106,7 @@ class MasterSite extends TimberSite
         add_action('init', [$this, 'p4_master_theme_setup']);
         add_action('pre_insert_term', [$this, 'disallow_insert_term'], 1, 2);
         add_filter('wp_dropdown_users_args', [$this, 'filter_authors'], 10, 1);
-        add_filter('wp_image_editors', [$this, 'allowedEditors']);
-        add_filter('wp_handle_upload_prefilter', [$this, 'image_type_validation']);
-        add_filter('jpeg_quality', fn () => 60);
         add_filter('http_request_timeout', fn () => 10);
-        add_action('after_setup_theme', [$this, 'add_image_sizes']);
         add_action('save_post', [$this, 'p4_auto_generate_excerpt'], 10, 2);
 
         add_action('admin_head', [$this, 'add_help_sidebar']);
@@ -172,10 +161,6 @@ class MasterSite extends TimberSite
             10,
             2
         );
-
-        version_compare(get_bloginfo('version'), '5.5', '<')
-            ? add_action('init', [$this, 'p4_register_core_image_block'])
-            : add_filter('register_block_type_args', [$this, 'register_core_blocks_callback']);
 
         // Disable WordPress(WP5.5) Block Directory.
         $this->disable_block_directory();
@@ -258,8 +243,6 @@ class MasterSite extends TimberSite
             10,
             1
         );
-
-        $this->register_meta_fields();
 
         // Override category template to use taxonomy.php.
         add_filter(
@@ -355,42 +338,6 @@ class MasterSite extends TimberSite
                 return $classes;
             }
         );
-    }
-
-    /**
-     * Sets as featured image of the post the first image found attached in the post's content (if any).
-     *
-     * @param int     $post_id The ID of the current Post.
-     * @param WP_Post $post The current Post.
-     */
-    public function set_featured_image(int $post_id, WP_Post $post): void
-    {
-        $types = Search\Filters\ContentTypes::get_all();
-        // Ignore autosave, check user's capabilities and post type.
-        if (
-            defined('DOING_AUTOSAVE') && DOING_AUTOSAVE
-            || !current_user_can('edit_post', $post_id)
-            || !in_array($post->post_type, array_keys($types))
-        ) {
-            return;
-        }
-
-        // Check if user has set the featured image manually.
-        $user_set_featured_image = get_post_meta($post_id, '_thumbnail_id', true);
-
-        // Apply this behavior only if there is not already a featured image.
-        if ($user_set_featured_image) {
-            return;
-        }
-
-        // Find all matches of <img> html tags within the post's content
-        // and get the id of the image from the elements class name.
-        preg_match_all('/<img.+wp-image-(\d+).*>/i', $post->post_content, $matches);
-        if (!isset($matches[1][0]) || !is_numeric($matches[1][0])) {
-            return;
-        }
-
-        set_post_thumbnail($post_id, $matches[1][0]);
     }
 
     /**
@@ -528,44 +475,6 @@ class MasterSite extends TimberSite
             $tab_post_url = get_permalink($tab_post_id);
             $nginx_purger->purge_url(user_trailingslashit($tab_post_url));
         }
-    }
-
-    /**
-     * Add extra image sizes as needed.
-     */
-    public function add_image_sizes(): void
-    {
-        add_image_size('retina-large', 2048, 1366, false);
-        add_image_size('articles-medium-large', 510, 340, false);
-    }
-
-    /**
-     * Force WordPress to use ImageCompression as image manipulation editor.
-     */
-    public function allowedEditors(): array
-    {
-        return [ImageCompression::class];
-    }
-
-    /**
-     * Validate immage type before WP processes it.
-     * * @param array $file Associative array containing Image details
-     */
-    public function image_type_validation(array $file): array
-    {
-        // Only apply validation to images.
-        $file_type = wp_check_filetype_and_ext($file['tmp_name'], $file['name']);
-        if (strpos($file_type['type'], 'image') !== 0) {
-            return $file;
-        }
-
-        $allowed_wp_img_ext = array('jpg', 'jpeg', 'png', 'gif', 'ico');
-
-        if (!in_array(strtolower($file_type['ext']), $allowed_wp_img_ext)) {
-            $file['error'] = 'Only JPG, PNG, ICO, and GIF images are allowed for upload.';
-        }
-
-        return $file;
     }
 
     /**
@@ -1058,23 +967,6 @@ class MasterSite extends TimberSite
     }
 
     /**
-     * Declare meta fields
-     */
-    private function register_meta_fields(): void
-    {
-        // Credit for images, used in image caption.
-        \register_post_meta(
-            'attachment',
-            self::CREDIT_META_FIELD,
-            [
-                'show_in_rest' => true,
-                'type' => 'string',
-                'single' => true,
-            ]
-        );
-    }
-
-    /**
      * Registers oembed provider for Carto map.
      */
     public function register_oembed_provider(): void
@@ -1163,83 +1055,6 @@ class MasterSite extends TimberSite
         $sidebar .= '<p><a target="_blank" href="https://planet4.greenpeace.org/">Planet 4 Handbook</a></p>';
 
         $screen->set_help_sidebar($sidebar);
-    }
-
-    /**
-     * Override the Gutenberg core/image block render method output,
-     * to add credit field in it's caption text & image alt text as title.
-     *
-     * @param array  $attributes    Attributes of the Gutenberg core/image block.
-     * @param string $content The image element HTML.
-     *
-     * @return string HTML content of image element with credit field in caption and alt text in image title.
-     */
-    public function p4_core_image_block_render(array $attributes, string $content): string
-    {
-        $image_id = isset($attributes['id']) ? trim(str_replace('attachment_', '', $attributes['id'])) : '';
-        $img_post_meta = $image_id ? get_post_meta($image_id) : [];
-        if (!$img_post_meta) {
-            return $content;
-        }
-
-        $credit = $img_post_meta[self::CREDIT_META_FIELD][0] ?? '';
-        $alt_text = $img_post_meta['_wp_attachment_image_alt'][0] ?? '';
-
-        if ($alt_text) {
-            $content = str_replace(' alt=', ' title="' . esc_attr($alt_text) . '" alt=', $content);
-        }
-
-        $image_credit = ' ' . $credit;
-        if (false === strpos($credit, '©')) {
-            $image_credit = ' ©' . $image_credit;
-        }
-
-        $caption = '';
-
-        $pattern = '/<figcaption[^>]*>(.*?)<\/figcaption>/';
-        if (preg_match($pattern, $content, $matches)) {
-            $caption = $matches[1];
-        }
-
-        if (empty($credit) || (!empty($caption) && strpos($caption, $image_credit) !== false)) {
-            return $content;
-        }
-
-        return str_replace(
-            empty($caption) ? '</figure>' : $caption,
-            empty($caption) ?
-                '<figcaption>' . esc_attr($image_credit) . '</figcaption></figure>' :
-                $caption . esc_attr($image_credit),
-            $content
-        );
-    }
-
-    /**
-     * Add callback function to Gutenberg core/image block.
-     */
-    public function p4_register_core_image_block(): void
-    {
-        unregister_block_type('core/image');
-        register_block_type(
-            'core/image',
-            ['render_callback' => [$this, 'p4_core_image_block_render']]
-        );
-    }
-
-    /**
-     * Add callback function to Gutenberg core/image block.
-     *
-     * @param array $args Parameters given during block register.
-     *
-     * @return array Parameters of the block.
-     */
-    public function register_core_blocks_callback(array $args): array
-    {
-        if ('core/image' === $args['name']) {
-            $args['render_callback'] = [$this, 'p4_core_image_block_render'];
-        }
-
-        return $args;
     }
 
     /**
